requirements.txt
Faker==24.1.2
joblib==1.3.2
numpy==1.26.2
pandas==2.1.2
plotly==5.18.0
plotly-express==0.4.1
pulp==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
scikit-learn==1.3.2
sentence-transformers==2.2.2
streamlit==1.28.2
torch==2.1.0
transformers==4.34.1
tensorflow==2.14.0
